refactor(enforce file name): Factor filter methods out of arfsdao PE-558
